Add calculator buttons
Added rows to html and added buttons to each row corresponding to the
buttons in a standard calculator. Adjusted styling of the rows and
containers so that they would be spaced correctly like a calculator.
Modified the querySelector in the js file so the selector corresponds
to the updated id.