Add button display funcs
Added id to main display of calculator for targeting with styling.
Altered styling of display container so that the content would wrap
instead of expanding the width of the calculator. Added functions
to the buttons in the js file so that the display will update each
time a button is pressed with its corresponding value