Add proper logic
Adjusted the logic in the arithmetic functions so that the numbers
returned would be correct from the operation and so that they would
round to the nearest second decimal place if necessary. Adjusted
the equals button so that it would call the evaluation and proper
arithmetic functions and then reset the values after completion. Wrote
the getEvaluation and resetValues functions to work properly when
called