Modify fringe logic bugs
Changed the divide function to make sure it checks for any value that
cannot be the divisor and alerts the user. Adjusted clear button so
that it actually clears the values on the backend. Adjusted operator
buttons to check if there were any operators already present so that the
arithmetic functions could be stringed together properly. Adjusted
getNum functions to limit the max characters to 9 in order to prevent
massive overflow. Modified getEvaluation to check for any erroneous
inputs and change the evaluation to a value that makes sense in the
situation presented to the function